Import of BBC/Pod-Xhtml-1.52 from CPAN.
gitpan-cpan-distribution: Pod-Xhtml
gitpan-cpan-version:      1.52
gitpan-cpan-path:         BBC/Pod-Xhtml-1.52.tar.gz
gitpan-cpan-author:       BBC
gitpan-cpan-maturity:     released

+++ t/01defaults.t
@@ -1,11 +1,12 @@
 #!/usr/bin/perl -w
-#$Id: 01defaults.t,v 1.19 2006/04/07 14:14:39 mattheww Exp $
+#$Id: 01defaults.t,v 1.21 2006/07/12 12:00:39 mattheww Exp $
 
 use strict;
 use lib qw(./lib ../lib);
 use Test;
 use Pod::Xhtml;
 use Getopt::Std;
+use File::Basename;
 
 getopts('tTs', \my %opt);
 if ($opt{t} || $opt{T}) {
@@ -13,18 +14,20 @@ if ($opt{t} || $opt{T}) {
 	import Log::Trace print => {Deep => $opt{T}};
 }
 
-if (-d 't') {
-	chdir( 't' );
-}
+chdir ( dirname ( $0 ) );
+
 require Test_LinkParser;
 
 my $filecont;
+my $goodcont;
+
 my $podia = 'a.pod';
-my $podoa = 'a.xhtml';
+my $podoa = 'a.pod.xhtml';
+my $podga = 'a.xhtml';
 
 unlink $podoa if -e $podoa;
 
-plan tests => 13;
+plan tests => 17;
 
 ok( $Pod::Xhtml::VERSION );
 
@@ -37,10 +40,12 @@ $parser->parse_from_file( $podia, $podoa );
 ok( -f $podoa );
 
 $filecont = readfile( $podoa );
+$goodcont = readfile( $podga );
+DUMP("filecont", \$filecont);
 ok( $filecont );
-ok( index( $filecont, cont_a() ) > -1 );
+ok( $filecont =~ m/\Q$goodcont\E/ );
 undef $filecont;
-unlink $podoa;
+unlink $podoa unless $opt{s};
 
 #### parsing from filehandles
 ok( ! -f $podoa );
@@ -50,195 +55,47 @@ close OUT;
 ok( -f $podoa );
 
 $filecont = readfile( $podoa );
+DUMP("filecont", \$filecont);
 ok( $filecont );
-ok( index( $filecont, cont_a() ) > -1 );
+ok( $filecont =~ m/\Q$goodcont\E/ );
 undef $filecont;
+undef $goodcont;
 unlink $podoa unless $opt{'s'};
 
 my $podib = 'b.pod';
-my $podob = 'b.xhtml';
-
+my $podob = 'b.pod.xhtml';
+my $podgb = 'b.xhtml';
+unlink $podob if -e $podob;
 ok ( !-f $podob );
 $parser->parse_from_file( $podib, $podob );
 ok ( -f $podob );
 
 $filecont = readfile( $podob );
+$goodcont = readfile( $podgb );
+DUMP("filecont", \$filecont);
 ok( $filecont );
-ok( index( $filecont, cont_b() ) > -1 );
+ok( $filecont =~ m/\Q$goodcont\E/ );
 undef $filecont;
+undef $goodcont;
 unlink $podob unless $opt{'s'};
 
-sub cont_a {
-return q(
-<body>
-<div class="pod">
-<!-- INDEX START -->
-<h3 id="TOP">Index</h3>
-<ul>
-	<li><a href="#NAME">NAME</a></li>
-	<li><a href="#SYNOPSIS">SYNOPSIS</a></li>
-	<li><a href="#DESCRIPTION">DESCRIPTION</a></li>
-	<li><a href="#Sed_diam_nomumny">Sed diam nomumny</a></li>
-	<li><a href="#METHODS">METHODS</a></li>
-	<li><a href="#ATTRIBUTES">ATTRIBUTES</a></li>
-	<li><a href="#ISSUES">ISSUES</a><br />
-<ul>
-	<li><a href="#KNOWN_ISSUES">KNOWN ISSUES</a></li>
-	<li><a href="#UNKNOWN_ISSUES">UNKNOWN ISSUES</a></li>
-</ul>
-</li>
-</ul>
-<hr />
-<!-- INDEX END -->
+my $podic = "c.pod";
+my $podoc = "c.pod.xhtml";
+my $podgc = "c.xhtml";
+unlink $podoc if -e $podoc;
+ok ( ! -f $podoc );
+$parser = Pod::Xhtml->new(LinkParser => $pod_links, MakeIndex => 2);
+$parser->parse_from_file( $podic, $podoc );
+ok ( -f $podoc );
 
-<h1 id="NAME">NAME</h1><p><a href="#TOP" class="toplink">Top</a></p>
-
-<p>A - Some demo POD</p>
-<h1 id="SYNOPSIS">SYNOPSIS</h1><p><a href="#TOP" class="toplink">Top</a></p>
-
-<pre>	use Pod::Xhtml;
-	my $px = new Pod::Xhtml;
-
-</pre>
-<h1 id="DESCRIPTION">DESCRIPTION</h1><p><a href="#TOP" class="toplink">Top</a></p>
-
-<p>This is a module to translate POD to Xhtml. Lorem ipsum <b>Dolor</b> in <cite>Dolor</cite> sit amet consectueur adipscing elit. Sed diam nomumny.
-This is a module to translate POD to Xhtml. <a href="#Lorem">The Lorem entry</a> ipsum dolor sit amet
-consectueur adipscing elit. Sed diam nomumny.
-This is a module to translate <cite>POD</cite> to Xhtml. <strong>Lorem</strong> ipsum <i>dolor</i> sit amet
-<code>consectueur adipscing</code> elit. <span id="Sed_diam_nomumny">Sed diam nomumny</span>.
-This is a module to translate POD to Xhtml. See <a href="#Lorem">Lorem</a> ipsum dolor sit amet
-consectueur adipscing elit. Sed diam <cite>nomumny</cite>. <a href="http://foo.bar/baz/">http://foo.bar/baz/</a></p>
-<h1 id="METHODS">METHODS</h1><p><a href="#TOP" class="toplink">Top</a></p>
-
-<dl>
-	<dt>Nested blocks</dt>
-	<dd>
-		<p>Pod::Xhtml now supports nested over/item/back blocks:</p>
-		<p>
-			<ul>
-					<li>Point 1				</li>
-					<li>Point Number 2				</li>
-					<li>Item three				</li>
-					<li>Point four
-<br /><br />Still point four
-<br /><br /><pre>  This is verbatim text in a bulleted list
-
-</pre></li>
-</ul>
-
-		</p>
-<pre>  This is verbatim test in a regular list
-
-</pre>
-	</dd>
-</dl>
-<h1 id="ATTRIBUTES">ATTRIBUTES</h1><p><a href="#TOP" class="toplink">Top</a></p>
-
-<dl>
-	<dt>Lorem</dt>
-	<dd>
-		<p>Lorem ipsum dolor sit amet consectueur adipscing elit. Sed diam nomumny.</p>
-	</dd>
-	<dt>Ipsum</dt>
-	<dd>
-		<p>Lorem ipsum dolor sit amet consectueur adipscing elit. Sed diam nomumny.</p>
-	</dd>
-	<dt>Dolor( $foo )</dt>
-	<dd>
-		<p>Lorem ipsum dolor sit amet consectueur .... elit. Sed diam nomumny.</p>
-	</dd>
-</dl>
-<h1 id="ISSUES">ISSUES</h1><p><a href="#TOP" class="toplink">Top</a></p>
-
-<h2 id="KNOWN_ISSUES">KNOWN ISSUES</h2>
-
-<p>There are some issues known about. Lorem ipsum dolor sit amet consectueur adipscing elit. Sed diam nomumny.
-Lorem ipsum dolor sit amet consectueur adipscing elit. Sed diam nomumny. SPACES&nbsp;&nbsp;&nbsp;ARE&nbsp;&nbsp;IMPORTANT</p>
-<h2 id="UNKNOWN_ISSUES">UNKNOWN ISSUES</h2>
-
-<p>There are also some issues not known about. Lorem ipsum dolor sit amet consectueur adipscing elit. Sed diam nomumny.
-Lorem ipsum dolor sit amet consectueur adipscing elit. Sed diam nomumny.</p>
-
-</div></body>
-);
-}
-
-sub cont_b {
-return q{
-<body>
-<div class="pod">
-<!-- INDEX START -->
-<h3 id="TOP">Index</h3>
-<ul>
-	<li><a href="#NAME">NAME</a></li>
-	<li><a href="#SYNOPSIS">SYNOPSIS</a><br />
-<ul>
-	<li><a href="#SUB_SYNOPSIS">SUB-SYNOPSIS</a></li>
-</ul>
-</li>
-	<li><a href="#DESCRIPTION">DESCRIPTION</a></li>
-	<li><a href="#LINKS">LINKS</a></li>
-	<li><a href="#ISSUES">ISSUES</a><br />
-<ul>
-	<li><a href="#KNOWN_ISSUES">KNOWN ISSUES</a><br />
-<ul>
-	<li><a href="#ARGV">$ARGV</a></li>
-	<li><a href="#ARGV-2">@ARGV</a></li>
-	<li><a href="#ARGV-3">%ARGV</a></li>
-	<li><a href="#Test_for_Escaped_HTML_in_Marked_text">Test for Escaped HTML in Marked text</a></li>
-</ul>
-</li>
-</ul>
-</li>
-</ul>
-<hr />
-<!-- INDEX END -->
-
-<h1 id="NAME">NAME</h1><p><a href="#TOP" class="toplink">Top</a></p>
-
-<p>B - Some demo POD</p>
-<h1 id="SYNOPSIS">SYNOPSIS</h1><p><a href="#TOP" class="toplink">Top</a></p>
-
-<pre>	use Pod::Xhtml;
-	my $px = new Pod::Xhtml;
-
-</pre>
-<h2 id="SUB_SYNOPSIS">SUB-SYNOPSIS</h2>
-
-<p>To test returning back to head1.</p>
-<h1 id="DESCRIPTION">DESCRIPTION</h1><p><a href="#TOP" class="toplink">Top</a></p>
-
-<p>This is a module to translate POD to Xhtml. Lorem ipsum <b>Dolor</b> in <cite>Dolor</cite> sit amet consectueur adipscing elit. Sed diam nomumny.</p>
-<h1 id="LINKS">LINKS</h1><p><a href="#TOP" class="toplink">Top</a></p>
-
-<p><a href="#ARGV-2">@ARGV</a> should link to the as-yet undefined &quot;<i>@ARGV</i>&quot; section</p>
-<p>Whereas <a href="#ARGV">$ARGV</a> shouldn't. It should link to the undefined
-&quot;<i>$ARGV</i>&quot; section</p>
-<h1 id="ISSUES">ISSUES</h1><p><a href="#TOP" class="toplink">Top</a></p>
-
-<h2 id="KNOWN_ISSUES">KNOWN ISSUES</h2>
-
-<h3 id="ARGV">$ARGV</h3>
-
-<p>Is sometimes undefined</p>
-<h3 id="ARGV-2">@ARGV</h3>
-
-<p>Is occasionally populated with the numbers 1, 2, 3, 4, 5, 6, 7, 8, 9 and 10</p>
-<h3 id="ARGV-3">%ARGV</h3>
-
-<p>Does not exist</p>
-<h3 id="Test_for_Escaped_HTML_in_Marked_text">Test for Escaped HTML in Marked text</h3>
-
-<p><code>&lt;meta /&gt;</code></p>
-<p><strong>R&amp;R</strong></p>
-<p><i>&quot;hello&quot;</i></p>
-
-</div></body>
-};
-}
-
-
+$filecont = readfile( $podoc );
+$goodcont = readfile( $podgc );
+DUMP("filecont", \$filecont);
+ok( $filecont );
+ok( $filecont =~ m/\Q$goodcont\E/ );
+undef $filecont;
+undef $goodcont;
+unlink $podoc unless $opt{'s'};
 
 sub readfile {
 	my $filename = shift;
@@ -311,6 +168,15 @@ Still point four
 
 =back
 
+=head2 TOP
+
+This should NOT reference #TOP, unless the top of the page has had its id
+changed, somehow, for some reason.
+
+=head2 EXAMPLE
+
+This is the first example block.
+
 =head1 ATTRIBUTES
 
 =over 4
@@ -325,9 +191,13 @@ Lorem ipsum dolor sit amet consectueur adipscing elit. Sed diam nomumny.
 
 =item Dolor( $foo )
 
-Lorem ipsum dolor sit amet consectueur ..Z<adipscing>.. elit. Sed diam nomumny.
+Lorem ipsum dolor sit amet consectueur ..Z<>.. elit. Sed diam nomumny.
 
 =back
+
+=head2 EXAMPLE
+
+This is the second example block.
 
 =head1 ISSUES
 
@@ -340,6 +210,10 @@ Lorem ipsum dolor sit amet consectueur adipscing elit. Sed diam nomumny. S<SPACE
 
 There are also some issues not known about. Lorem ipsum dolor sit amet consectueur adipscing elit. Sed diam nomumny.
 Lorem ipsum dolor sit amet consectueur adipscing elit. Sed diam nomumny.
+
+=head3 EXAMPLE
+
+This is the third example block.
 
 =cut
